+++ FE/src/page/forecast.html
@@ -23,13 +23,7 @@
 
       <main id="dust__forecast">
         <h1 class="dust__title">미세먼지 예보</h1>
-        <div class="forecast__map">
-          <!-- <img
-            class="map__img"
-            src="https://search.pstatic.net/common?quality=75&direct=true&src=https%3A%2F%2Fcsearch-phinf.pstatic.net%2F20200330_30%2F1585555694758cwG0C_PNG%2Fimage.png"
-            alt="미세먼지 예보 이미지"
-          /> -->
-        </div>
+        <div class="forecast__map"></div>
         <div class="forecast__container">
           <p class="forecast__content">
             미세먼지 영상은 환경부 국립과학원이 에어코리아를 통해 발표하는
@@ -63,6 +57,6 @@
         </div>
       </main>
     </div>
-    <script src="../js/forecast.js" type="module"></script>
+    <script src="../js/forecast/main.js" type="module"></script>
   </body>
 </html>
